changed names to. make things more clear, added seperate commands to make things easier
