+++ 2026-2706-Robot-Code/src/main/java/frc/robot/commands/IntakeDownCommand.java
@@ -3,27 +3,22 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.IntakeSubsystem;
 
-public class RunIntakeCommand extends Command {
+public class IntakeDownCommand extends Command {
 
     private final IntakeSubsystem intake;
 
-    public RunIntakeCommand(IntakeSubsystem intake) {
+    public IntakeDownCommand(IntakeSubsystem intake) {
         this.intake = intake;
         addRequirements(intake);
     }
 
     @Override
     public void initialize() {
-        intake.startIntake();
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        intake.stopIntake();
+        intake.moveIntakeDown();
     }
 
     @Override
     public boolean isFinished() {
-        return false;
+        return true;
     }
 }
