+++ 2026-2706-Robot-Code/src/main/java/frc/robot/commands/RunIntakeCommandForward.java
@@ -1,0 +1,29 @@
+package frc.robot.commands;
+
+import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.subsystems.IntakeSubsystem;
+
+public class RunIntakeCommandForward extends Command {
+
+    private final IntakeSubsystem intake;
+
+    public RunIntakeCommandForward(IntakeSubsystem intake) {
+        this.intake = intake;
+        addRequirements(intake);
+    }
+
+    @Override
+    public void initialize() {
+        intake.startIntake();
+    }
+
+    @Override
+    public void end(boolean interrupted) {
+        intake.stopIntake();
+    }
+
+    @Override
+    public boolean isFinished() {
+        return false;
+    }
+}
